# docs/openapi/bap-v1.yaml
openapi: 3.0.3
info:
  title: Blueprint Agent Protocol (BAP) v1
  version: "2.0.0"
  description: |
    The API surface external autonomous agents (e.g. Hermes) use to operate
    Blueprint without the web dashboard. Mounted at `/api/bap/v1`.

    ## Authentication

    Every route except `POST /register` and `GET /discover` requires a
    `BAP-Key` header (or `Authorization: Bearer <key>`) with a key issued
    by `POST /register`. Keys look like `bap_<36 random base64url chars>`.

    `POST /register` itself requires **either** an authenticated dashboard
    session cookie, **or** a shared `X-Registration-Secret` header matching
    the operator-configured `BAP_REGISTRATION_SECRET` environment variable.
    It is never open to unauthenticated callers.

    Three `GET /agents-admin/*` routes use the dashboard's own session
    cookie instead of `BAP-Key` — they're operator-only and are documented
    separately at the bottom of this spec with the `sessionAuth` scheme.

    ## Idempotency

    Every mutating endpoint (`POST`/`PATCH` that creates or changes state)
    **requires** an `Idempotency-Key` header. Reusing the same key with an
    identical request body replays the original response instead of
    re-running the handler; reusing it with a different body returns `409`.
    A request with no `Idempotency-Key` header on a route that requires one
    returns `400`. See the `IdempotencyKey` parameter and `Conflict`
...

    ## Pagination

...

    ## Timestamps

...

...

...

...

...

  contact:
...
servers:
...
security:
  - bapKey: []

# ─────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bapKey:
      type: apiKey
      in: header
      name: BAP-Key
      description: A key issued by POST /register. Authorization Bearer <key> is also accepted.
    registrationSecret:
      type: apiKey
      in: header
      name: X-Registration-Secret
      description: Shared secret configured via BAP_REGISTRATION_SECRET. Only required if not using a dashboard session.
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Dashboard session cookie — operator-only routes (/agents-admin/*), not part of the external-agent BAP surface.
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
      description: Required on every mutating endpoint. See the top-level Idempotency section.
...
    BusinessIdPath:
      name: businessId
      in: path
      required: true
      schema: { type: string }
  responses:
    NotFound:
      description: The resource does not exist.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: "Task 'tsk_xxx' not found.", code: not_found, request_id: "..." }
    PermissionDenied:
      description: The caller's key lacks the required permission, or the resource belongs to a business outside the key's business_access.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: "Permission denied: task does not belong to an authorized business.", code: permission_denied, request_id: "..." }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string, enum: [validation_error, not_found, permission_denied, conflict, rate_limited, internal_error] }
        request_id: { type: string, nullable: true }
        retryable: { type: boolean }
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        pages: { type: integer }
...
    Recommendation:
      type: object
      description: "getRankedRecommendations() output, enriched with an explanation object per item (server/routes/bap-review.ts explainRecommendation())."
      properties:
        id: { type: string }
        source_type: { type: string, enum: [task, opportunity, strategy] }
        title: { type: string }
        score: { type: number }
        rationale: { type: array, items: { type: string } }
        expected_impact: { type: string, nullable: true }
        confidence: { type: number, nullable: true }
        estimated_cost: { type: number, nullable: true }
        estimated_effort: { type: string, nullable: true }
        historical_success_rate: { type: number, nullable: true }
        historical_sample_size: { type: integer }
        goal_id: { type: string, nullable: true }
        goal_title: { type: string, nullable: true }
        has_open_conflict: { type: boolean }
        explanation:
          type: object
          properties:
            evidence: { type: array, items: { type: string } }
            reasoning: { type: array, items: { type: string } }
            alternatives_considered: { type: array, items: { type: object } }
            confidence: { type: number, nullable: true }
            risk: { type: string }
            expected_impact: { type: string, nullable: true }
            historical_evidence:
              type: object
              properties:
                success_rate: { type: number, nullable: true }
                sample_size: { type: integer }
            linked_kb: { type: array, items: { type: object }, description: "Only populated for the top 5 (explanation_depth) — [] otherwise, not missing/omitted." }
            linked_signals: { type: array, items: { type: object } }
            linked_metrics: { type: object, nullable: true }
    ExcludedRecommendation:
      type: object
      properties:
        id: { type: string }
        source_type: { type: string, enum: [task, opportunity, strategy] }
        title: { type: string }
        reason: { type: string, description: "The blocking constraint's note, or a generic fallback." }
    RetrospectiveSummary:
      type: object
      properties:
        id: { type: string }
        business_id: { type: string }
        period_start: { type: string, format: date-time }
        period_end: { type: string, format: date-time }
        executive_summary: { type: string, nullable: true }
        kb_path: { type: string, nullable: true }
        triggered_by: { type: string, default: scheduler }
        created_at: { type: string, format: date-time }
    RetrospectiveDetail:
      allOf:
        - { $ref: '#/components/schemas/RetrospectiveSummary' }
        - type: object
          properties:
            what_worked: { type: array, items: { type: object } }
            what_didnt: { type: array, items: { type: object } }
            learnings: { type: array, items: { type: object } }
            agent_assessments: { type: array, items: { type: object } }
            open_windows: { type: array, items: { type: object } }
            recommendations: { type: array, items: { type: object } }
            operating_changes: { type: array, items: { type: object } }
            calibration_notes: { type: array, items: { type: object } }
            full_report_json: { type: object, nullable: true }
    CalibrationRow:
      type: object
      properties:
        id: { type: string }
        agent_id: { type: string }
        business_id: { type: string, nullable: true }
        period_start: { type: string, format: date-time }
        period_end: { type: string, format: date-time }
        tasks_with_outcomes: { type: integer }
        avg_stated_confidence: { type: number, nullable: true }
        avg_actual_outcome_rate: { type: number, nullable: true }
        calibration_error: { type: number, nullable: true }
        calibration_score: { type: number, nullable: true }
        calibration_offset: { type: number, default: 0 }
        trend: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        false_positives: { type: integer, default: 0 }
        false_negatives: { type: integer, default: 0 }
        recommendations_accepted: { type: integer, default: 0 }
        recommendations_rejected: { type: integer, default: 0 }
        execution_success_rate: { type: number, nullable: true }
        long_term_success_rate: { type: number, nullable: true }
        calibration_method: { type: string, default: simple_average, description: "'simple_average' marks rows computed before Phase 3's richer calibration existed." }
        conservatism_factor: { type: number, default: 1.0 }
        bins: { type: array, items: { type: object } }
        calculated_at: { type: string, format: date-time }
    CrossBusinessPattern:
      type: object
      description: "No business_id, name, URL, or task title — enforced structurally, the underlying table has no such column."
      properties:
        pattern_key: { type: string }
        action_type: { type: string }
        description: { type: string }
        sample_size: { type: integer }
        success_count: { type: integer }
        success_rate: { type: number, nullable: true }
        applicable_business_types: { type: array, items: { type: string }, description: "Business types (e.g. ecommerce, saas), never names." }
...
paths:
...
  /businesses/{businessId}/recommendations:
    get:
      operationId: listRecommendations
      summary: Ranked, explainable recommendations across tasks/opportunities/strategies
      x-bap-permission: recommendations:read
      parameters:
        - { $ref: '#/components/parameters/BusinessIdPath' }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  recommendations: { type: array, items: { $ref: '#/components/schemas/Recommendation' } }
                  excluded: { type: array, items: { $ref: '#/components/schemas/ExcludedRecommendation' } }
                  explanation_depth: { type: string, description: "e.g. 'full for the top 5, partial (no KB search) for the rest'." }
                  total: { type: integer }
        '404': { $ref: '#/components/responses/NotFound' }
  /businesses/{businessId}/retrospectives:
    get:
      operationId: listRetrospectives
      summary: List retrospectives, newest first, paginated
      x-bap-permission: retrospectives:read
      parameters:
        - { $ref: '#/components/parameters/BusinessIdPath' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  retrospectives: { type: array, items: { $ref: '#/components/schemas/RetrospectiveSummary' } }
                  total: { type: integer }
                  pagination: { $ref: '#/components/schemas/Pagination' }
  /retrospectives/{retrospectiveId}:
    get:
      operationId: getRetrospectiveDetail
      summary: Retrospective detail — full report
      x-bap-permission: retrospectives:read
      parameters: [{ name: retrospectiveId, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  retrospective: { $ref: '#/components/schemas/RetrospectiveDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
  /businesses/{businessId}/retrospectives/run:
    post:
      operationId: runRetrospective
      summary: Trigger a retrospective now
      x-bap-permission: retrospectives:trigger
      description: >-
        Fire-and-forget — an LLM call over a month of data plus a KB write.
        Poll GET /businesses/{businessId}/retrospectives for the result.
      parameters:
        - { $ref: '#/components/parameters/BusinessIdPath' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '202': { description: "{ business_id, status: 'running', message }" }
        '404': { $ref: '#/components/responses/NotFound' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
  /businesses/{businessId}/calibration:
    get:
      operationId: getCalibration
      summary: Agent calibration — latest per agent, or full history for one agent
      x-bap-permission: calibration:read
      parameters:
        - { $ref: '#/components/parameters/BusinessIdPath' }
        - { name: agent_id, in: query, schema: { type: string } }
        - { name: history, in: query, schema: { type: string }, description: "Any truthy value returns calibration_history (up to 50 rows) instead of the latest-per-agent list." }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      calibration: { type: array, items: { $ref: '#/components/schemas/CalibrationRow' } }
                      total: { type: integer }
                  - type: object
                    properties:
                      calibration_history: { type: array, items: { $ref: '#/components/schemas/CalibrationRow' } }
  /businesses/{businessId}/patterns:
    get:
      operationId: listCrossBusinessPatterns
      summary: Cross-business learning — abstracted patterns only, no tenant-identifying data
      x-bap-permission: recommendations:read
      description: "No separate patterns:read permission — reuses recommendations:read."
      parameters: [{ $ref: '#/components/parameters/BusinessIdPath' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  patterns: { type: array, items: { $ref: '#/components/schemas/CrossBusinessPattern' } }
                  total: { type: integer }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Phase 2: Connectors ─────────────────────────────────────────────────
...